Merge on Lucas Branch

+++ src/main/java/br/edu/ifsp/domain/usecases/ativo/fundodeinvestimento/ExcluirFundoDeInvestimentoUseCase.java
@@ -3,6 +3,7 @@ package br.edu.ifsp.domain.usecases.ativo.fundodeinvestimento;
 import br.edu.ifsp.domain.entities.ativo.FundoDeInvestimento;
 import br.edu.ifsp.domain.entities.log.LogAtivo;
 import br.edu.ifsp.domain.entities.log.LogAtivoEnum;
+import br.edu.ifsp.domain.usecases.grupo.GrupoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.LogAtivoDAO;
 import br.edu.ifsp.domain.usecases.log.logativo.SalvarHistoricoAtivoUseCase;
 import br.edu.ifsp.domain.usecases.utils.Notification;
@@ -11,10 +12,12 @@ import br.edu.ifsp.domain.usecases.utils.Validator;
 public class ExcluirFundoDeInvestimentoUseCase {
     private FundoDeInvestimentoDAO fundoDeInvestimentoDAO;
     private LogAtivoDAO logAtivoDAO;
+    private GrupoDAO grupoDAO;
 
-    public ExcluirFundoDeInvestimentoUseCase(FundoDeInvestimentoDAO fundoDeInvestimentoDAO, LogAtivoDAO logAtivoDAO) {
+    public ExcluirFundoDeInvestimentoUseCase(FundoDeInvestimentoDAO fundoDeInvestimentoDAO, LogAtivoDAO logAtivoDAO, GrupoDAO grupoDAO) {
         this.fundoDeInvestimentoDAO = fundoDeInvestimentoDAO;
         this.logAtivoDAO = logAtivoDAO;
+        this.grupoDAO = grupoDAO;
     }
 
     public boolean delete(FundoDeInvestimento fundoDeInvestimento) {
@@ -23,6 +26,10 @@ public class ExcluirFundoDeInvestimentoUseCase {
 
         if(notif.hasErrors()) {
             throw new IllegalArgumentException(notif.errorMessage());
+        }
+
+        if(this.grupoDAO.findOneByAtivo(fundoDeInvestimento).isPresent()) {
+            throw new IllegalArgumentException("Cant delete ativo thats in a grupo");
         }
 
         boolean flag = this.fundoDeInvestimentoDAO.delete(fundoDeInvestimento);
